requirements.txt
aerich==0.8.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncclick==8.1.8.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
descope==1.7.4
dictdiffer==0.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
iso8601==2.1.0
liccheck==0.9.2
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.10.1
pypika-tortoise==0.5.0
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
semantic-version==2.10.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.2
toml==0.10.2
tomlkit==0.13.2
tortoise-orm==0.25.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
